Major refactor: split up frontend stuff into different apps as well, create core django app and have other apps share its templates, tell webpack to package common stuff like react in separate bundles
