+++ class/assets/bundles/.gitignore
@@ -1,0 +1,2 @@
+# keep the assets/bundles directory in git, but nothing in it
+*
