+++ class/core/__init__.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class CoreApp(AppConfig):
+    name = 'core'
