+++ class/frontends/reimburse/index.js
@@ -3,12 +3,10 @@ var ReactDOM = require('react-dom')
 
 var Hello = React.createClass({
     render: function() {
-	return (
-	    <h1>
-		Hello, React!
-	    </h1>
-	)	
+        return (
+            <h1>Hello, React!</h1>
+        )
     }    
 })
 
-ReactDOM.render(<Hello />, document.getElementById('container'))
+ReactDOM.render(<Hello />, document.getElementById('reactbase'))
