+++ class/frontends/webpack-stats.json
@@ -1,0 +1,1 @@
+{"status":"done","chunks":{"reimburse":[{"name":"reimburse.js","path":"/home/dk-proj/class-3.0/class/assets/bundles/reimburse.js"}],"commons":[{"name":"commons.js","path":"/home/dk-proj/class-3.0/class/assets/bundles/commons.js"}]}}
