--- class/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "class",
-  "version": "1.0.0",
-  "description": "",
-  "main": "webpack.config.js",
-  "scripts": {
-      "build" : "webpack --config webpack.config.js --watch"
-  },
-  "author": "",
-  "license": "ISC"
-}
