+++ class/reimburse/__init__.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ReimburseApp(AppConfig):
+    name = 'reimburse'
