--- class/reimburse/templates/reimburse/index.html
@@ -1,10 +1,0 @@
-{% load render_bundle from webpack_loader %}
-<html>
-    <head>
-	<title> DLP: Reimburse </title>
-    </head>
-    <body>
-	<div id="container"></div>
-	{% render_bundle 'main' %}
-    </body>
-</html>
