+++ class/reimburse/views.py
@@ -1,6 +1,9 @@
 from django.http import HttpResponse
 from django.template import loader
 
+
 def index(request):
-    template = loader.get_template('reimburse/index.html')
-    return HttpResponse(template.render({}, request))
+    template = loader.get_template('core/index.html')
+    return HttpResponse(template.render({
+        'bundle': 'reimburse'
+    }, request))
