--- class/webpack-stats.json
@@ -1,1 +1,0 @@
-{"status":"done","chunks":{"main":[{"name":"main-e4e99023cff05a7f5ee6.js","path":"C:\\Users\\Brian\\Documents\\harvard\\DLP\\class-3.0\\class\\assets\\bundles\\main-e4e99023cff05a7f5ee6.js"}]}}
